Synchronization experiments.

+++ audio/ffmpeg.h
@@ -4,7 +4,7 @@
 
 /// Generic audio decoder (using ffmpeg)
 struct AudioFile {
-	uint channels = 0;
+	size_t channels = 0;
 	uint audioFrameRate = 0;
 	int64 audioTime = 0;
     uint duration = 0;
